Move info area into individual namespace.
